refactor: add event rules file
