+++ src/shield/rules/group.rules.ts
@@ -1,7 +1,6 @@
 import { rule } from "graphql-shield";
 import { UNAUTHORIZED } from "../../common/common.constants";
 import { Context } from "../../context/context.service";
-import { Event } from "../../events/models/event.model";
 import { GroupPrivacy } from "../../groups/group-configs/models/group-config.model";
 import { UpdateGroupConfigInput } from "../../groups/group-configs/models/update-group-config.input";
 import { CreateGroupRoleInput } from "../../groups/group-roles/models/create-group-role.input";
@@ -199,31 +198,6 @@ export const isPublicGroupImage = rule()(
       "group.config",
     ]);
     return image?.group?.config.privacy === GroupPrivacy.Public;
-  }
-);
-
-export const isPublicEventImage = rule()(
-  async (parent, _args, { services: { imagesService } }: Context) => {
-    const image = await imagesService.getImage({ id: parent.id }, [
-      "event.group.config",
-    ]);
-    return image?.event?.group?.config.privacy === GroupPrivacy.Public;
-  }
-);
-
-export const isPublicGroupEvent = rule()(
-  async (
-    parent: Event,
-    args: { id: number },
-    { services: { eventsService } }: Context
-  ) => {
-    const event = await eventsService.getEvent({
-      id: parent ? parent.id : args.id,
-      group: {
-        config: { privacy: GroupPrivacy.Public },
-      },
-    });
-    return !!event;
   }
 );
